block: Add driver to change the I/O scheduler when the screen turns off
This driver changes the I/O scheduler for a specified request queue to
Noop when the screen turns off in order to improve power consumption while
the screen is off, as well as reduce resume latency when the screen is
turned back on. After the screen is turned back on, the driver will change
the I/O scheduler back to what it originally was after a 10000ms delay.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>
(cherry picked from commit 87f935c831f52be0aeeb0609b180bf2bfb1ec3fa)
Signed-off-by: Enzaik <imernestodiaz@gmail.com>